+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
+    "dotenv": "^8.0.0",
     "express": "~4.16.1",
     "mongoose": "^5.6.7",
     "morgan": "~1.9.1"
