+++ routes/api/index.js
@@ -1,6 +1,7 @@
 var router = require('express').Router()
 
-router.use('/api', require('./api'))
+router.use('users', require('users'))
+/* GET home page. */
 // router.get('/', function (req, res, next) {
 //   res.render('index', { title: 'Express' })
 // })
